add str_bsearch.c & 諸々の例外処理の細かいところを直した

+++ includes/free_func.h
@@ -6,7 +6,7 @@
 /*   By: tjinichi <tjinichi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/11 21:18:22 by tjinichi          #+#    #+#             */
-/*   Updated: 2020/12/11 21:18:29 by tjinichi         ###   ########.fr       */
+/*   Updated: 2020/12/17 22:31:56 by tjinichi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # include <stdlib.h>
 
-void	*ptr_2d_free(void ***ptr, int cur);
+void	*ptr_2d_free(void ***ptr, size_t cur);
 void	ptr_free(void **ptr);
 
 #endif
